Silence compiler warnings by using designated datatypes for size and offset

+++ include/dom/core/characterdata.h
@@ -25,21 +25,21 @@ typedef struct dom_characterdata_vtable {
 			dom_string *data);
 	dom_exception (*dom_characterdata_get_length)(
 			struct dom_characterdata *cdata,
-			uint32_t *length);
+			size_t *length);
 	dom_exception (*dom_characterdata_substring_data)(
-			struct dom_characterdata *cdata, uint32_t offset,
+			struct dom_characterdata *cdata, off_t offset,
 			uint32_t count, dom_string **data);
 	dom_exception (*dom_characterdata_append_data)(
 			struct dom_characterdata *cdata,
 			dom_string *data);
 	dom_exception (*dom_characterdata_insert_data)(
 			struct dom_characterdata *cdata,
-			uint32_t offset, dom_string *data);
+			off_t offset, dom_string *data);
 	dom_exception (*dom_characterdata_delete_data)(
 			struct dom_characterdata *cdata,
-			uint32_t offset, uint32_t count);
+			off_t offset, uint32_t count);
 	dom_exception (*dom_characterdata_replace_data)(
-			struct dom_characterdata *cdata, uint32_t offset,
+			struct dom_characterdata *cdata, off_t offset,
 			uint32_t count, dom_string *data);
 } dom_characterdata_vtable;
 
@@ -63,16 +63,16 @@ static inline dom_exception dom_characterdata_set_data(
 		(struct dom_characterdata *) (c), (d))
 
 static inline dom_exception dom_characterdata_get_length(
-		struct dom_characterdata *cdata, uint32_t *length)
+		struct dom_characterdata *cdata, size_t *length)
 {
 	return ((dom_characterdata_vtable *) ((dom_node *) cdata)->vtable)->
 			dom_characterdata_get_length(cdata, length);
 }
 #define dom_characterdata_get_length(c, l) dom_characterdata_get_length( \
-		(struct dom_characterdata *) (c), (uint32_t *) (l))
+		(struct dom_characterdata *) (c), (size_t *) (l))
 
 static inline dom_exception dom_characterdata_substring_data(
-		struct dom_characterdata *cdata, uint32_t offset,
+		struct dom_characterdata *cdata, off_t offset,
 		uint32_t count, dom_string **data)
 {
 	return ((dom_characterdata_vtable *) ((dom_node *) cdata)->vtable)->
@@ -81,7 +81,7 @@ static inline dom_exception dom_characterdata_substring_data(
 }
 #define dom_characterdata_substring_data(c, o, ct, d) \
 		dom_characterdata_substring_data( \
-		(struct dom_characterdata *) (c), (uint32_t) (o), \
+		(struct dom_characterdata *) (c), (off_t) (o), \
 		(uint32_t) (ct), (d))
 
 static inline dom_exception dom_characterdata_append_data(
@@ -94,28 +94,28 @@ static inline dom_exception dom_characterdata_append_data(
 		(struct dom_characterdata *) (c), (d))
 
 static inline dom_exception dom_characterdata_insert_data(
-		struct dom_characterdata *cdata, uint32_t offset, 
+		struct dom_characterdata *cdata, off_t offset,
 		dom_string *data)
 {
 	return ((dom_characterdata_vtable *) ((dom_node *) cdata)->vtable)->
 			dom_characterdata_insert_data(cdata, offset, data);
 }
 #define dom_characterdata_insert_data(c, o, d) dom_characterdata_insert_data( \
-		(struct dom_characterdata *) (c), (uint32_t) (o), (d))
+		(struct dom_characterdata *) (c), (off_t) (o), (d))
 
 static inline dom_exception dom_characterdata_delete_data(
-		struct dom_characterdata *cdata, uint32_t offset, 
+		struct dom_characterdata *cdata, off_t offset,
 		uint32_t count)
 {
 	return ((dom_characterdata_vtable *) ((dom_node *) cdata)->vtable)->
 			dom_characterdata_delete_data(cdata, offset, count);
 }
 #define dom_characterdata_delete_data(c, o, ct) dom_characterdata_delete_data(\
-		(struct dom_characterdata *) (c), (uint32_t) (o), \
+		(struct dom_characterdata *) (c), (off_t) (o), \
 		(uint32_t) (ct))
 
 static inline dom_exception dom_characterdata_replace_data(
-		struct dom_characterdata *cdata, uint32_t offset, 
+		struct dom_characterdata *cdata, off_t offset,
 		uint32_t count, dom_string *data)
 {
 	return ((dom_characterdata_vtable *) ((dom_node *) cdata)->vtable)->
@@ -124,7 +124,7 @@ static inline dom_exception dom_characterdata_replace_data(
 }
 #define dom_characterdata_replace_data(c, o, ct, d) \
 		dom_characterdata_replace_data(\
-		(struct dom_characterdata *) (c), (uint32_t) (o),\
+		(struct dom_characterdata *) (c), (off_t) (o),\
 		(uint32_t) (ct), (d))
 
 #endif
